Add handle errors configuration.

+++ src/main/java/com/btg/funds/service/FundService.java
@@ -1,5 +1,7 @@
 package com.btg.funds.service;
 
+import com.btg.funds.exceptions.FundNotFoundException;
+import com.btg.funds.exceptions.InsufficientFundsException;
 import com.btg.funds.model.Fund;
 import com.btg.funds.model.Transaction;
 import com.btg.funds.repository.FundRepository;
@@ -17,7 +19,6 @@ public class FundService {
 
     private static final double INITIAL_BALANCE = 500_000;
 
-
     @Autowired
     private FundRepository fundRepository;
 
@@ -31,12 +32,14 @@ public class FundService {
 
     private final String emailToCatch = "ashmeddiazg@gmail.com";
 
-
     public Optional<Fund> getFundById(Integer id) {
         return fundRepository.findById(id);
     }
 
     public void cancelFund(Integer id) {
+        if (!fundRepository.existsById(id)) {
+            throw new FundNotFoundException("Fondo con ID " + id + " no encontrado");
+        }
         fundRepository.deleteById(id);
     }
 
@@ -46,24 +49,22 @@ public class FundService {
 
     public String subscribeFund(Fund fund) {
         if (userBalance < fund.getMinAmount()) {
-            notificationService.sendNotification(emailToCatch,"No tiene saldo disponible para vincularse al fondo " + fund.getName());
-            return "No tiene saldo disponible para vincularse al fondo " + fund.getName();
+            String message = "No tiene saldo disponible para vincularse al fondo " + fund.getName();
+            notificationService.sendNotification(emailToCatch, message);
+            throw new InsufficientFundsException(message);
         }
 
         userBalance -= fund.getMinAmount();
         Transaction transaction = createTransaction(fund, "APERTURA");
         transactionRepository.save(transaction);
-        notificationService.sendNotification(emailToCatch,"Suscripción realizada exitosamente al fondo " + fund.getName());
+        notificationService.sendNotification(emailToCatch, "Suscripción realizada exitosamente al fondo " + fund.getName());
         return "Suscripción realizada exitosamente al fondo " + fund.getName();
     }
 
     public String unsubscribeFund(Integer fundId) {
-        Optional<Fund> fundOptional = fundRepository.findById(fundId);
-        if (fundOptional.isEmpty()) {
-            return "Fondo no encontrado";
-        }
+        Fund fund = fundRepository.findById(fundId)
+                .orElseThrow(() -> new FundNotFoundException("Fondo con ID " + fundId + " no encontrado"));
 
-        Fund fund = fundOptional.get();
         userBalance += fund.getMinAmount();
         Transaction transaction = createTransaction(fund, "CANCELACION");
         transactionRepository.save(transaction);
